object_recognition/assignment3/requirements.txt
matplotlib==3.4.2
numpy==1.21.4
Pillow==8.4.0
skimage==0.0
torch==1.10.0+cu113
torchsummary==1.5.1
torchvision==0.11.1+cu113
tqdm==4.62.3
